solves issue with Int valueTypes that could not be used.
